Puesta a punto de detalles post arreglos

+++ Logic/GamesLogic.cs
@@ -60,9 +60,16 @@ namespace Logic
             return ret;
         }
 
-        public void Delete(int gameId)
+        public string Delete(int gameId)
         {
+            var game = GetById(gameId);
+            if (game == null)
+            {
+                return "El juego ingresado no existe en el sistema";
+            }
+            
             _gamesRepository.Delete(gameId);
+            return $"Game with id {gameId} was deleted from the store.";
         }
 
         public List<Game> GetPublishedGames(User userLogged)
